+++ app/scripts/controllers/kwic_controller.js
@@ -2,7 +2,6 @@
 import _ from "lodash"
 import settings from "@/settings"
 import model from "@/model"
-import { prettyNumbers } from "@/util"
 
 const korpApp = angular.module("korpApp")
 
@@ -47,6 +46,10 @@ export class KwicCtrl {
         const s = scope
 
         s.initialSearch = true
+        /** Number of total search hits, updated when a search is completed. */
+        s.hits = undefined
+        /** Number of search hits, may change while search is in progress. */
+        s.hitsProgress = undefined
 
         this.setupListeners()
 
@@ -147,7 +150,7 @@ export class KwicCtrl {
         s.onProgress = (progressObj, isPaging) => {
             s.progress = Math.round(progressObj["stats"])
             if (!isPaging && progressObj["total_results"] !== null) {
-                s.hits_display = prettyNumbers(progressObj["total_results"])
+                s.hitsInProgress = progressObj["total_results"]
             }
         }
 
@@ -210,7 +213,7 @@ export class KwicCtrl {
             s.loading = false
             if (!isPaging) {
                 s.hits = data.hits
-                s.hits_display = prettyNumbers(data.hits)
+                s.hitsInProgress = data.hits
                 s.corpusHits = data.corpus_hits
             }
         }